Release 2.0.0 hardening: ML-DSA docs, MSRV CI, and expanded test coverage (#7)
* feat: add prehash and expand tests

* ci: add GitHub Actions pipeline and cargo-deny policy

* refactor: format codebase and clean up tests/benches

* refactor: resolve clippy warnings and preserve bit-packing semantics

* ci: speed up audit tools install and add scheduled security scan

* chore(deps): update criterion and lockfile to fix cargo-deny advisory

* ci: run only lib and tests in test job (exclude benches)

* refactor(security,api): zeroize secret keys and make key byte fields private

* chore(deny): allow Unicode-DFS-2016 license for transitive deps

* feat!: harden key APIs with Result, bump to v2.0.0, add rustdoc gate in CI

* feat!: tighten public API and finalize ML-DSA/FIPS docs

* ci(msrv): enforce Rust 1.78 for library build

* test(ml-dsa): add negative, boundary, and serialization coverage for 44/65/87

* update README

---------

+++ src/reduce.rs
@@ -6,7 +6,7 @@ pub const Q_INV: i32 = 58728449; // q^(-1) mod 2^32
 /// Returns r.
 pub fn montgomery_reduce(a: i64) -> i32 {
     let mut t = (a as i32).wrapping_mul(Q_INV) as i64;
-    t = (a as i64 - t.wrapping_mul(crate::params::Q as i64)) >> 32;
+    t = (a - t.wrapping_mul(crate::params::Q as i64)) >> 32;
     t as i32
 }
